requirements.txt
appdirs
black
certifi
chardet
click
dataclasses
Flask
future
idna
itsdangerous
Jinja2
MarkupSafe
mypy-extensions
pathspec
pyadaptivecards
PyJWT
python-dotenv
regex
requests
requests-toolbelt
tabulate
toml
typed-ast
typing-extensions
urllib3
webexteamsbot
webexteamssdk
Werkzeug
google-api-python-client
pandas